authorization system updates (#2868)
* first commit


* debug


* authz api first try


* service api done


* change authz check order to avoid panic


* change authorization logic to avoid possible panic


* build module done & merge user client for picket(aslan)


* fix policy(aslan)-user dependency


* delivery module done


* moving code


* debug


* update collaboration mode apis


* env half done


* environment part fully done


* project done


* remove unused package


* debug logs & some system level authorization problems


* fix project list logic


* added query logic in mongodb


* fix collaboration mode creation error


* system authorization done


* collaboration mode fix for environment


* more collaboration mode fix for environments


* minor debug


* delivery center authz change


* workflow view authz fix


* workflow v4 filter & trigger filter


* debug logs for list workflows


* workflow module done for test


* fix init_info api authz problem


* fix production env creation problem


* fixed envcfg related api authz


* multiple authorization bugfix


* remove admin initialization from aslan, moving it to user


* remove useless code


* user db moved


* testing module done


* debug


* debug


* debug


* fix list test workflows


* added todo in comment


* system config related authorization done


* minor bugfix


* fix workflow plugin stuff


* user related api update


* workflow edit bug fix


* debug for list workflow task in product workflow


* added collaboration mode permission check in user and use it in aslan


* build service module authorization fix


* added list authorized env api to make life easier & multiple environment api authorization fix


* fix multiple permission problem with comments


* extend permission for multiple apis temporarily


* debug


* remove project-only authorization


* get user role update


* revert api deletion


* read-project-only permission change


* update collaboration mode handle logic


* bugfix


* debug info & warn log removal


* even more debug logs


* environment filter update


* fixed wrong list env logic


* fix panic bug


* added testing filtering


* debug logs & added project auth info for public project


* new permission logic for menu


* remove useless code


* remove bundle


* added some policies back


* remove useless code


* implement multiple authorization leaks on purpose


* remove authorization check for helm chart version


* debug


* update production env support for user


* multiple auth leaks


* debug


* production env permission


* ...'


* bugfix for permission


* debug update service


* debug


* remove useless comments


---------

+++ pkg/microservice/aslan/core/system/handler/announcement.go
@@ -17,8 +17,9 @@ limitations under the License.
 package handler
 
 import (
-	"github.com/gin-gonic/gin"
+	"fmt"
 
+	"github.com/gin-gonic/gin"
 	systemmodel "github.com/koderover/zadig/pkg/microservice/aslan/core/system/repository/models"
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/system/service"
 	internalhandler "github.com/koderover/zadig/pkg/shared/handler"
@@ -26,28 +27,58 @@ import (
 )
 
 func CreateAnnouncement(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
+
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
+
 	args := new(systemmodel.Announcement)
 
-	err := c.BindJSON(args)
+	err = c.BindJSON(args)
 	if err != nil {
 		ctx.Err = e.ErrInvalidParam.AddDesc("invalid notify args")
 		return
 	}
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
+
 	ctx.Err = service.CreateAnnouncement(ctx.UserName, args, ctx.Logger)
 }
 
 func UpdateAnnouncement(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
+
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
+
 	args := new(systemmodel.Announcement)
 
-	err := c.BindJSON(args)
+	err = c.BindJSON(args)
 	if err != nil {
 		ctx.Err = e.ErrInvalidParam.AddDesc("invalid notify args")
 		return
 	}
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
+
 	ctx.Err = service.UpdateAnnouncement(ctx.UserName, args.ID.Hex(), args, ctx.Logger)
 }
 
@@ -66,10 +97,23 @@ func PullNotifyAnnouncement(c *gin.Context) {
 }
 
 func DeleteAnnouncement(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
+	ctx, err := internalhandler.NewContextWithAuthorization(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
 
+	if err != nil {
+		ctx.Logger.Errorf("failed to generate authorization info for user: %s, error: %s", ctx.UserID, err)
+		ctx.Err = fmt.Errorf("authorization Info Generation failed: err %s", err)
+		ctx.UnAuthorized = true
+		return
+	}
+
 	ID := c.Param("id")
+
+	// authorization checks
+	if !ctx.Resources.IsSystemAdmin {
+		ctx.UnAuthorized = true
+		return
+	}
 
 	ctx.Err = service.DeleteAnnouncement(ctx.UserName, ID, ctx.Logger)
 }
